requirements.txt
asgiref==3.8.1
Django==5.1.1
django-appconf==1.0.6
django-bootstrap5==24.2
django-environ==0.11.2
django-imagekit==5.0.0
pilkit==3.0
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
